Test for invalid repo_depends items (closes #1414)
Use newer Python as the original version of this patch uses lilac, and
the latter has `from __future__ import annotations`, which is not
available until Python 3.7. Now newer Python is not neccessary, but it
is still better to upgrade sooner than later.

And repo_depends are now denied in lilac.py (#990)